shared/tls: Move from shared
